+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 This is a roll-up of all release notes in order of release
 
+## [Release 3.2.0 - December 2, 2015](https://github.com/KrisSiegel/msngr.js/releases/tag/3.2.0)
+- Fixed a bug in ```msngr.querySelectorAllWithEq()``` where more than one :eq() would create an invalid selector
+- Revised ```msngr.getDomPath() ```to properly return a DOM path instead of just adding an ID. If an ID exists that is simply returned as the selector
+- Revised ```msngr.findElement()``` / ```msngr.findElements()``` to use querySelectorAllWithEq() for consistency
+- Moved test resources to a dedicated folder for neatness
+
 ## [Release 3.1.1 - November 15, 2015](https://github.com/KrisSiegel/msngr.js/releases/tag/3.1.1)
 - Heavily revised README
 - Added CHANGELOG for tracking release information